[add] add a lot 基本完成目前所有功能，待调试

+++ earlywarning-machine/src/main/java/com/alexzfx/earlywarningmachine/rabbit/MachineMessage.java
@@ -12,9 +12,9 @@ import java.io.Serializable;
 @Data
 public class MachineMessage implements Serializable {
     int machineId;
-    int data;
+    double data;
 
-    public MachineMessage(int machineId, int data) {
+    public MachineMessage(int machineId, double data) {
         this.machineId = machineId;
         this.data = data;
     }
